third_party/blink/web_tests/external/wpt/css/css-values/acos-asin-atan-atan2-computed-expected.txt

This is a testharness.js-based test.
FAIL acos(1) should be used-value-equivalent to 0deg assert_equals: rotate(acos(1)) and rotate(0deg) serialize to the same thing in used values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL atan(0) should be used-value-equivalent to 0deg assert_equals: rotate(atan(0)) and rotate(0deg) serialize to the same thing in used values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL asin(0) should be used-value-equivalent to 0deg assert_equals: rotate(asin(0)) and rotate(0deg) serialize to the same thing in used values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL atan2(0,0) should be used-value-equivalent to 0deg assert_equals: rotate(atan2(0,0)) and rotate(0deg) serialize to the same thing in used values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL calc(asin(sin(pi/2))) should be used-value-equivalent to 90deg Cannot read properties of undefined (reading 'split')
FAIL calc(acos(cos(pi - 3.14159265358979323846))) should be used-value-equivalent to 0deg Cannot read properties of undefined (reading 'split')
FAIL calc(atan(e - 2.7182818284590452354) ) should be used-value-equivalent to 0deg Cannot read properties of undefined (reading 'split')
FAIL calc(asin(sin(30deg + 1.0471967rad ) )) should be used-value-equivalent to 90deg Cannot read properties of undefined (reading 'split')
PASS calc(acos(cos(30deg - 0.523599rad ) )) should be used-value-equivalent to 0deg
FAIL calc(asin(sin(3.14159 / 2 + 1 - 1) )) should be used-value-equivalent to 90deg Cannot read properties of undefined (reading 'split')
FAIL calc(asin(sin(100grad) )) should be used-value-equivalent to 90deg Cannot read properties of undefined (reading 'split')
FAIL calc(acos(cos(0 / 2 + 1 - 1) )) should be used-value-equivalent to 0deg Cannot read properties of undefined (reading 'split')
FAIL calc(atan(tan(30deg + 0.261799rad ) )) should be used-value-equivalent to 45deg Cannot read properties of undefined (reading 'split')
FAIL calc(atan(tan(0.7853975rad ) )) should be used-value-equivalent to 45deg Cannot read properties of undefined (reading 'split')
FAIL calc(atan(tan(3.14159 / 4 + 1 - 1) )) should be used-value-equivalent to 45deg Cannot read properties of undefined (reading 'split')
FAIL calc(asin(sin(0.25turn)) ) should be used-value-equivalent to 90deg Cannot read properties of undefined (reading 'split')
FAIL calc(atan2(0,1)) should be used-value-equivalent to 0deg Cannot read properties of undefined (reading 'split')
FAIL calc(atan2(0,-1)) should be used-value-equivalent to -180deg Cannot read properties of undefined (reading 'split')
FAIL calc(atan2(1,-1)) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL calc(atan2(-1,1)) should be used-value-equivalent to -45deg Cannot read properties of undefined (reading 'split')
FAIL calc(cos(sin(acos(cos(pi))))) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(sin(atan(tan(pi/2)))) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL atan2(1px, -1px) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1cm, -1cm) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1mm, -1mm) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1Q, -1Q) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1in, -1in) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1pc, -1pc) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1pt, -1pt) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1em, -1em) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1ex, -1ex) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1ch, -1ch) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1rem, -1rem) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1vh, -1vh) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1vw, -1vw) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1deg, -1deg) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1grad, -1grad) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1turn, -1turn) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1rad, -1rad) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1s, -1s) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
FAIL atan2(1ms, -1ms) should be used-value-equivalent to 135deg Cannot read properties of undefined (reading 'split')
Harness: the test ran to completion.
